analyze mcmc samplings
