+++ hq/cli/main.py
@@ -196,6 +196,42 @@ class CLI:
 
         sys.exit(0)
 
+    def analyze_mcmc(self):
+        """Analyze the samplings generated by running MCMC on unimodal stars"""
+        from .analyze_joker_samplings import analyze_joker_samplings
+        parser = get_parser(
+            description=(
+                "Generate statistics and metadata for the samplings "
+                "created by running The Joker."))
+        # HACK
+        parser.usage = 'hq analyze_thejoker' + parser.format_usage()[9:]
+
+        args = parser.parse_args(sys.argv[2:])
+
+        with threadpool_limits(limits=1, user_api='blas'):
+            with args.Pool(**args.Pool_kwargs) as pool:
+                analyze_joker_samplings(args.run_path, pool=pool)
+
+        sys.exit(0)
+
+    def analyze_mcmc(self):
+        """Analyze the samplings generated by MCMC"""
+        from .analyze_mcmc_samplings import analyze_mcmc_samplings
+        parser = get_parser(
+            description=(
+                "Generate statistics and metadata for the samplings "
+                "created by running MCMC."))
+        # HACK
+        parser.usage = 'hq analyze_mcmc' + parser.format_usage()[9:]
+
+        args = parser.parse_args(sys.argv[2:])
+
+        with threadpool_limits(limits=1, user_api='blas'):
+            with args.Pool(**args.Pool_kwargs) as pool:
+                analyze_mcmc_samplings(args.run_path, pool=pool)
+
+        sys.exit(0)
+
 
 # Auto-generate the usage block:
 cmds = []
